Add missing article
This turned out to probably not be a typo,
despite what I assumed when committing 79cbeaf.
In French, it's the exception rather than the rule
to ommit the article before a noun. One such exception
is with proper nouns (i.e. Euler), but not with the
noun 'formule'. If that had been a test, this would
have been one of those pesky trick questions.