Fix #10
Judging how appropriate any specific
use of the comma may be is tricky ---
trickier than I gave it credit for.
I'll try to keep an eye out for my
tendency to over-use the symbol, while
still making use of it to make sentences
clearer. In short, the use of a comma
before 'et' is frowned upon, but is also
justified and accepted in some particular
cases.