Fix spacing after mid-sentence punctuation
Similar to, inter alia, cd17847. By default, when LaTeX encounters a
period, question mark, exclamation point, or other punctuation which is
typically sentence-final immediately before a space, it makes the space
slightly larger than a regular interword space. If the punctuation is
instead being used in the middle of a sentence, as here, adding \ after
it (and before the space) prevents the space from being embiggened.