Replace 'on' by 'nous'
We have a strange relationship with 'on'. It's
a good replacement to 'nous' in a lot of cases,
but it's a bit of a houserule and not always a
good call. Having a possessive pronoun here makes
it tricky: in spoken French, we would probably say
'on commence nos...', but the 'nos' refers to 'on'
and, grammatically, they're not the same person.