Rephrase 'a place to point them to'
Choosing the more specific 'site' because
'indiquer un endroit' ('to point to a place')
was not a common phrase in this context. It
would work for a literal place, like a monument
or a restaurant. I don't know why I'm explaining
what a place is... I gotta stop drinking this much
coffee. Although, now that I think about it, I
_could_ go for one more cup. Ok, one last cup...