Rephrase 'less confusing'
'Confusing' is a concept wholly foreign
to the French. Indeed, I've never needed to
express the idea ever, except when I was
abroad. That, or we have convoluted ways to
say it, and we're too myopic to consciously
adapt. Anyway, I italicised the phrase,
rather than the adverb, to preserve the
symmetry with the question.