Rephrase 'I care about X of Y'
Tricky, that one. 'correctness of explanation' can,
and I believe, must, be simplified to 'correct explanations'.
I may be able to find a more poetic way to phrase it
that would still be fairly natural, but that's dubious
and I don't think it's necessary. This way, the
lyricism is reasonnably well preserved in the symmetry
and the progression of the ideas along with the wording.

'to care' as 'se soucier' would be ok if we had stuck
closer to the text. I think 'recherche' works fine
to describe something he wants to both obtain and provide.