Remove emphatic 'seem'
The phrase 'seemed to notice' seems to be an
emphatic form of 'noticed'. It provides rhythm,
gives more time to a potential speaker to go up the
scale for the final, comedic high-pitched indignation.
Or I'm reading into it too much. But I don't think so,
and I think the expletive 'ne' that you put there
already does a similar job.